# Use set in parser scope