--- a/packages/babel-parser/src/plugins/typescript/scope.js
+++ b/packages/babel-parser/src/plugins/typescript/scope.js
@@ -14,22 +14,22 @@ import {
 import * as N from "../../types";
 
 class TypeScriptScope extends Scope {
-  types: string[] = [];
+  types: Set<string> = new Set();
 
   // enums (which are also in .types)
-  enums: string[] = [];
+  enums: Set<string> = new Set();
 
   // const enums (which are also in .enums and .types)
-  constEnums: string[] = [];
+  constEnums: Set<string> = new Set();
 
   // classes (which are also in .lexical) and interface (which are also in .types)
-  classes: string[] = [];
+  classes: Set<string> = new Set();
 
   // namespaces and ambient functions (or classes) are too difficult to track,
   // especially without type analysis.
   // We need to track them anyway, to avoid "X is not defined" errors
   // when exporting them.
-  exportOnlyBindings: string[] = [];
+  exportOnlyBindings: Set<string> = new Set();
 }
 
 // See https://github.com/babel/babel/pull/9766#discussion_r268920730 for an
@@ -44,7 +44,7 @@ export default class TypeScriptScopeHandler extends ScopeHandler<TypeScriptScope
     const scope = this.currentScope();
     if (bindingType & BIND_FLAGS_TS_EXPORT_ONLY) {
       this.maybeExportDefined(scope, name);
-      scope.exportOnlyBindings.push(name);
+      scope.exportOnlyBindings.add(name);
       return;
     }
 
@@ -56,11 +56,11 @@ export default class TypeScriptScopeHandler extends ScopeHandler<TypeScriptScope
         this.checkRedeclarationInScope(scope, name, bindingType, pos);
         this.maybeExportDefined(scope, name);
       }
-      scope.types.push(name);
+      scope.types.add(name);
     }
-    if (bindingType & BIND_FLAGS_TS_ENUM) scope.enums.push(name);
-    if (bindingType & BIND_FLAGS_TS_CONST_ENUM) scope.constEnums.push(name);
-    if (bindingType & BIND_FLAGS_CLASS) scope.classes.push(name);
+    if (bindingType & BIND_FLAGS_TS_ENUM) scope.enums.add(name);
+    if (bindingType & BIND_FLAGS_TS_CONST_ENUM) scope.constEnums.add(name);
+    if (bindingType & BIND_FLAGS_CLASS) scope.classes.add(name);
   }
 
   isRedeclaredInScope(
@@ -68,18 +68,18 @@ export default class TypeScriptScopeHandler extends ScopeHandler<TypeScriptScope
     name: string,
     bindingType: BindingTypes,
   ): boolean {
-    if (scope.enums.indexOf(name) > -1) {
+    if (scope.enums.has(name)) {
       if (bindingType & BIND_FLAGS_TS_ENUM) {
         // Enums can be merged with other enums if they are both
         //  const or both non-const.
         const isConst = !!(bindingType & BIND_FLAGS_TS_CONST_ENUM);
-        const wasConst = scope.constEnums.indexOf(name) > -1;
+        const wasConst = scope.constEnums.has(name);
         return isConst !== wasConst;
       }
       return true;
     }
-    if (bindingType & BIND_FLAGS_CLASS && scope.classes.indexOf(name) > -1) {
-      if (scope.lexical.indexOf(name) > -1) {
+    if (bindingType & BIND_FLAGS_CLASS && scope.classes.has(name)) {
+      if (scope.lexical.has(name)) {
         // Classes can be merged with interfaces
         return !!(bindingType & BIND_KIND_VALUE);
       } else {
@@ -87,7 +87,7 @@ export default class TypeScriptScopeHandler extends ScopeHandler<TypeScriptScope
         return false;
       }
     }
-    if (bindingType & BIND_KIND_TYPE && scope.types.indexOf(name) > -1) {
+    if (bindingType & BIND_KIND_TYPE && scope.types.has(name)) {
       return true;
     }
 
@@ -95,9 +95,11 @@ export default class TypeScriptScopeHandler extends ScopeHandler<TypeScriptScope
   }
 
   checkLocalExport(id: N.Identifier) {
+    const topLevelScope = this.scopeStack[0];
+    const { name } = id;
     if (
-      this.scopeStack[0].types.indexOf(id.name) === -1 &&
-      this.scopeStack[0].exportOnlyBindings.indexOf(id.name) === -1
+      !topLevelScope.types.has(name) &&
+      !topLevelScope.exportOnlyBindings.has(name)
     ) {
       super.checkLocalExport(id);
     }