… packages/babel-parser/benchmark/many-let-declarations-within-block/10-length-binding.mjs