<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Fixed Issues?            | Parsing let/const declarations costs O(n^2) time
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->
This PR is similar to #13406: We use Set for checking duplicate bindings in parser scope.

Benchmark result (baseline is 7.14.4, current is this PR)

<details>
  <summary>Length-1 let declarations</summary>

```
$ node --predictable ./many-let-declarations/1-length-binding.mjs

baseline 256 length-1 let bindings: 1772 ops/sec ±21.58% (0.564ms)
baseline 512 length-1 let bindings: 733 ops/sec ±0.68% (1.365ms)
baseline 1024 length-1 let bindings: 232 ops/sec ±0.78% (4.312ms)
baseline 2048 length-1 let bindings: 66.35 ops/sec ±1.49% (15ms)

current 256 length-1 let bindings: 2592 ops/sec ±26.52% (0.386ms)
current 512 length-1 let bindings: 1497 ops/sec ±1.32% (0.668ms)
current 1024 length-1 let bindings: 728 ops/sec ±0.85% (1.374ms)
current 2048 length-1 let bindings: 323 ops/sec ±1.32% (3.093ms)
```
</details>

<a href="https://gitpod.io/#https://github.com/babel/babel/pull/13408"><img src="https://gitpod.io/button/open-in-gitpod.svg"/></a>