@motiz88 2021-06-01
Just in case this isn't already on your radar @JLHwung, I recently noticed something similar in `@babel/traverse`'s scope tracking - e.g. `referencedPaths` should be a Set in https://github.com/babel/babel/blob/b281fe352c6ed2f984613aa3b1a2da66bce2dafc/packages/babel-traverse/src/scope/binding.ts#L94-L101

(Of course, there it's technically an API change so may not be straightforward to just change, but maybe we can add a secondary set to speed up checking for duplicates while still maintaining the array for compatibility.)