@JLHwung 2021-06-01
@motiz88 Thanks. That's a good catch! 

A secondary set may introduce consistent issues due to multiple truth. We could provide `referencePaths` as an accessor for `referencePathSet` and migrate built-in plugins first.